// src/app/model/control-template-data-model.ts
export class ControlTemplateDataModel {
    name: string;
    type: string;
    label: string;
    value: any;
    required: boolean;
    adminOnly: boolean;
    helpText: string;
    showHelpText: boolean;
    showToolTip: boolean;
    showMyLabel: boolean;
    content: boolean;
    options: string[];
}